==> experiments_paper/step1_Simulation.py <==

import jpype
import numpy as np
from jpype import *
from jpype.types import *
import jpype.imports
import json
import pickle

from AutoCD.data_utils.function_simulate_tigramite import *
from AutoCD.data_utils.functions_data import *
from AutoCD.causal_graph_utils.markov_boundary import *
from AutoCD.causal_graph_utils.create_sub_mag_pag import *


# ----------------------------------------------------------------------
# Experiment: Simulate temporal data using Tigramite and Tetrad projects
# Author: kbiza@csd.uoc.gr
# ----------------------------------------------------------------------

def main():

    jpype.startJVM("-ea", classpath=['../jar_files/*'], convertStrings=False)


    # Simulation parameters
    params={}
    params['n_rep'] = 2  # number of repetitions
    params['n_nodes'] = 10  # number of temporal variables
    params['n_samples'] = 1500  # number of samples
    params['n_train'] = 500    # number of samples to hold for analysis
    params['avg_degree'] = 2   # average node degree in each time-lag
    params['max_degree'] = 4   # maximum node degree in each time-lag
    params['n_lags'] = 2       # maximum number of time lags

    params['is_time_series'] = True


    # File names
    path = './files_results/'
    id_file = (str(params['n_nodes'] ) + 'n_' +
               str(params['avg_degree'])+'d_' +
               str(params['n_lags']) + 'lags_' +
               str(params['n_samples']) + 's_')
    files_name = path + id_file + 'files.pkl'
    params_name = path + id_file + 'params.pkl'


    # Simulation
    files = [{} for _ in range(params['n_rep'])]
    rep = 0
    while rep < params['n_rep']:

        print('\n\n REP:', rep)
        flag_data = True
        while flag_data:
            print('Try to simulate data')
            _data_syn, dag_pd, tigramite_graph, links = (
                simulate_time_series(params['n_nodes'],
                                     params['n_samples'],
                                     params['n_lags'],
                                     params['avg_degree'],
                                     params['max_degree'],
                                     seed=None))
            if dag_pd.shape[1] < params['n_nodes'] * (params['n_lags']+1):
                print('smaller time lagged graph')
            elif np.any(np.abs(np.max(_data_syn)>10^3)):
                print('large numbers')
            else:
                flag_data = False


        # Target
        flag_target = True
        while flag_target:
            target_idx = np.random.randint(1, params['n_nodes'])
            target_name = 'V'+str(target_idx+1)
            if np.count_nonzero(dag_pd.iloc[:,target_idx]==2) > 2:
                flag_target = False


        # True Mb
        true_dag = dag_pd.copy()
        true_mb_idx = markov_boundary(target_idx, true_dag)
        true_mb_nameslag = true_dag.columns[true_mb_idx]
        true_mb_names = names_from_lag(true_mb_nameslag)

        # True PAG using true MB
        all_vars = true_dag.columns
        study_vars_true = np.concatenate((true_mb_names, [target_name]))
        study_vars_true = list(set(study_vars_true))
        _, true_pag_ = create_sub_mag_pag(true_dag, study_vars_true, params['n_lags'])

        true_pag_study = pd.DataFrame(np.zeros((len(all_vars), len(all_vars)), dtype=int),columns=all_vars, index=all_vars)
        true_pag_study.loc[true_pag_.index, true_pag_.columns] = true_pag_.loc[true_pag_.index, true_pag_.columns]


        # Transform data
        data_train = _data_syn.iloc[0:params['n_train'],:].copy()
        data_test = _data_syn.iloc[params['n_train']:-1, :].copy()

        data_type_train = get_data_type(data_train)
        data_type_test = get_data_type(data_test)
        transformed_train = transform_data(data_train, data_type_train, 'standardize')
        transformed_test = transform_data(data_test, data_type_test, 'standardize')

        # Based on the simulation function, we assume that a linear regression model is the true model
        true_pred_config={}
        true_pred_config['pred_name'] = 'linear_regression'

        # Save
        files[rep]['data'] = _data_syn
        files[rep]['data_train'] = transformed_train
        files[rep]['data_test'] = transformed_test
        files[rep]['target_name'] = target_name
        files[rep]['true_dag'] = dag_pd
        files[rep]['true_mb_names'] = true_mb_names
        files[rep]['true_mb_nameslag'] = true_mb_nameslag
        files[rep]['true_pag_study'] = true_pag_study
        files[rep]['true_pred_config'] = true_pred_config

        rep+=1

    with open(files_name, 'wb') as outp:
        pickle.dump(files, outp, pickle.HIGHEST_PROTOCOL)

    with open(params_name, 'wb') as outp:
        pickle.dump(params, outp, pickle.HIGHEST_PROTOCOL)


if __name__ == "__main__":
    main()
